refactor: slot 유틸리티를 _utils/slot으로 이동
