+++ packages/ui/src/_utils/slot/slot.tsx
@@ -4,5 +4,7 @@ type WithoutChildren = Omit<useRender.ComponentProps<'div'>, 'children'>;
 export interface SlotProps extends WithoutChildren {}
 
 export const Slot = ({ render, ...props }: SlotProps) => {
+  if (!render) return null;
+
   return useRender({ render, props });
 };
