+++ packages/ui/src/button/Button.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { createSlots, type SlotsProps } from '../slot';
+import { createSlots, type SlotsProps } from '../_utils/slot';
 import * as PrimitiveButton from './index.primitives';
 
 /* ------ slots ------ */
