+++ packages/ui/src/popover/popover.tsx
@@ -1,6 +1,6 @@
-import type { SlotsProps } from '../slot';
+import type { SlotsProps } from '../_utils/slot';
 
-import { createSlots } from '../slot';
+import { createSlots } from '../_utils/slot';
 import * as PrimitivePopover from './index.primitives';
 
 /* ----- slots ----- */
